backend/ai_agent/requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama
python-dotenv
httpx
pytest
